Fix timer UI update issue with type mismatch
- Change getIntExtra to getLongExtra in MainActivity for remainingTime
- Update TimerAction.updateTimer parameter from Int to Long
- Ensures proper data type consistency between service and UI

Fixes issue where timer UI was not showing remaining time updates.

+++ app/src/main/java/io/github/hanihashemi/tomaten/MainActivity.kt
@@ -103,7 +103,7 @@ class MainActivity : ComponentActivity() {
                     context: Context?,
                     intent: Intent?,
                 ) {
-                    val remainingTime = intent?.getIntExtra(REMAINING_TIME, 0) ?: 0
+                    val remainingTime = intent?.getLongExtra(REMAINING_TIME, 0) ?: 0
                     viewModel.actions.timer.updateTimer(remainingTime)
                 }
             }
